// Fis_Ama_EsgAnalytics/Fis_Ama_EsgAnalytics/ClientApp/src/app/models/constants/companyNameInStorage.ts
export enum companyNameInStorage {
  abbott = "companyData_abbott",
  novartis = "companyData_novartis",
  cipla = "companyData_cipla",
  apple = "companyData_apple",
  accenture = "companyData_accenture",
  amazon = "companyData_amazon",
  ford = "companyData_ford",
  tesla = "companyData_tesla",
  cocacola ="companyData_cocacola",
  goair = "companyData_goair",
  airasia = "companyData_airasia",
  vistara = "companyData_vistara"
}